adding a abstract base dataset class so dataset modules like the opnex module and the newly added procgen module can inherit from it
